backend/requirements.txt

fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai>=2.0.0
google-genai
faiss-cpu
pypdf
python-docx
pandas
python-dotenv
tiktoken
sentence-transformers
